+++ source/Private/Get-DscResourceSchemaPropertyContent.ps1
@@ -34,7 +34,11 @@ function Get-DscResourceSchemaPropertyContent
     (
         [Parameter(Mandatory = $true)]
         [System.Collections.Hashtable[]]
-        $Property
+        $Property,
+
+        [Parameter()]
+        [System.Management.Automation.SwitchParameter]
+        $UseMarkdown
     )
 
     $stringArray = [System.String[]] @()
@@ -76,7 +80,16 @@ function Get-DscResourceSchemaPropertyContent
 
         if (-not [System.String]::IsNullOrEmpty($currentProperty.ValueMap))
         {
-            $propertyLine += ' ' + ($currentProperty.ValueMap -join ', ')
+            $valueMap = $currentProperty.ValueMap
+
+            if ($UseMarkdown.IsPresent)
+            {
+                $valueMap = $valueMap | ForEach-Object -Process {
+                    '`{0}`' -f $_
+                }
+            }
+
+            $propertyLine += ' ' + ($valueMap -join ', ')
         }
 
         $propertyLine += ' |'